+++ shellcheck-repl.bash
@@ -9,9 +9,9 @@
 #'
 #' License: ISC
 #' Home page: https://github.com/HenrikBengtsson/shellcheck-repl
-#' Version: 0.1.4
+
 sc_repl_version() {
-    echo "0.1.4"
+    echo "0.2.0"
 }
 
 ## Source: https://github.com/koalaman/shellcheck/issues/1535
@@ -115,8 +115,15 @@ sc_repl_asserts() {
 }    
 
 sc_repl_verify_or_unbind() {
+    local input
+    local opts    
     local skip_pattern
-
+    local style
+    local start_time
+    local end_time
+    local vars
+    local tmp
+    
     sc_repl_debug "sc_repl_verify_or_unbind() ..."
     
     ## Skip ShellCheck? Default is to skip with leading:
@@ -124,12 +131,12 @@ sc_repl_verify_or_unbind() {
     ## * DOUBLESPACE$ (in-house rule)
     skip_pattern=${SHELLCHECK_REPL_SKIP_PATTERN:-(^\!|[[:space:]][[:space:]]$)}
     if [[ "$READLINE_LINE" =~  $skip_pattern ]]; then
-        sc_repl_debug " - skipping because it matches SHELLCHECK_REPL_SKIP_PATTERN=${SHELLCHECK_REPL_SKIP_PATTERN}"
+        sc_repl_debug " - skip pattern matched: ${skip_pattern}"
         sc_repl_debug "sc_repl_verify_or_unbind() ... done"
         return
     fi
     
-    local opts=("--shell=bash" "--external-sources")
+    opts=("--shell=bash" "--external-sources")
     if [[ -n "${SHELLCHECK_REPL_EXCLUDE}" ]]; then
         opts+=("--exclude=${SHELLCHECK_REPL_EXCLUDE}")
     fi
@@ -142,29 +149,61 @@ sc_repl_verify_or_unbind() {
         opts+=("--severity=${SHELLCHECK_REPL_VERIFY_LEVEL:=info}")
     fi
     # Filter the output of shellcheck by removing filename
-    local style=${SHELLCHECK_REPL_INFO,,}
+    style=${SHELLCHECK_REPL_INFO,,}
+    if [[ -z "${style}" ]]; then style="clean"; fi
     sc_repl_debug " - style: ${style}"
     sc_repl_debug " - ShellCheck options: ${opts[*]}"
-    if [[ -z "${style}" ]]; then style="clean"; fi
+    sc_repl_debug " - READLINE_LINE: ${READLINE_LINE}"
+
+    ## Identify variables to be included in preamble
+    vars=()
+    if [[ ! "$SHELLCHECK_REPL_EXCLUDE" == *2154* ]]; then
+        sc_repl_debug " - Special case: Checking with rule SC2154"
+        ## Ask shellcheck to identify variables of interest
+        mapfile -t vars < <(shellcheck --shell=bash --format=gcc --exclude="${SHELLCHECK_REPL_EXCLUDE}" <(echo "$READLINE_LINE") | grep -F "[SC2154]" | sed -E 's/.*warning: ([^ ]+) .*/\1/')
+    fi
+
+    if [[ ! "$SHELLCHECK_REPL_EXCLUDE" == *2178* ]]; then
+        sc_repl_debug " - Special case: Checking with rule SC2178"
+        ## Ask shellcheck to identify variables of interest
+        mapfile -t tmp < <(shellcheck --shell=bash --format=gcc <(echo "$READLINE_LINE") | grep -F "[SC2034]" | sed -E 's/.*warning: ([^ ]+) .*/\1/')
+        vars+=("${tmp[@]}")
+    fi
+    
+    sc_repl_debug " - Variables: [n=${#vars[@]}] ${vars[*]}"
+    
+    ## Are there any variables involved?
+    if [[ ${#vars[@]} -gt 0 ]]; then
+        ## 'declare -p' dump only those
+        input=$(declare -p "${vars[@]}" 2> /dev/null)
+        input=$(printf "#dummy to disable does not apply to everything\ntrue\n#shellcheck disable=all\n{\ntrue\n%s\n}\n\n%s\n" "${input}" "$READLINE_LINE")
+    else
+        input=$READLINE_LINE
+    fi
+    
+
+    sc_repl_debug " - ShellCheck input: $(echo "${input}" | wc -l) lines"
+    start_time=$(date +%s%N)    
     case ${style} in
         raw-tty)
-	    shellcheck "${opts[@]}" --format=tty <(printf '%s\n' "$READLINE_LINE")
+	    shellcheck "${opts[@]}" --format=tty <(echo "${input}")
 	    ;;
         raw-gcc)
-	    shellcheck "${opts[@]}" --format=gcc <(printf '%s\n' "$READLINE_LINE")
+	    shellcheck "${opts[@]}" --format=gcc <(echo "${input}")
 	    ;;
         full)
-	    shellcheck "${opts[@]}" <(printf '%s\n' "$READLINE_LINE") | tail -n +2
+	    shellcheck "${opts[@]}" <(echo "${input}") | tail -n +2
 	    ;;
         clean)
-	    shellcheck "${opts[@]}" <(printf '%s\n' "$READLINE_LINE") | sed -n '1,2b; /^$/q; p'
+	    shellcheck "${opts[@]}" <(echo "${input}") | sed -n '1,2b; /^$/q; p'
 	    ;;
         note)
-	    shellcheck "${opts[@]}" --format=gcc <(printf '%s\n' "$READLINE_LINE") | cut -d : -f 4- ;;
+	    shellcheck "${opts[@]}" --format=gcc <(echo "${input}") | cut -d : -f 4- ;;
 	*)
             sc_repl_error "Unknown value for shellcheck-repl variable 'SHELLCHECK_REPL_INFO' (valid values are 'raw', 'full', 'short' and 'clean' [default]): '${SHELLCHECK_REPL_INFO}'"
 	    ;;
     esac
+    
     if [[ "${PIPESTATUS[0]}" != 0 ]]; then
 	>&2 echo
 	>&2 echo "To skip a check, add its SC number to 'SHELLCHECK_REPL_EXCLUDE', e.g."
@@ -181,7 +220,11 @@ sc_repl_verify_or_unbind() {
         bind -x '"\C-x\C-b2": sc_repl_verify_bind_accept'
         sc_repl_assert_keybind_exists "\C-x\C-b2"
     fi
+
+    end_time=$(date +%s%N)
     
+    sc_repl_debug " - check time: $(((end_time - start_time) / 1000000)) ms"
+
     sc_repl_debug "sc_repl_verify_or_unbind() ... done"
 }
 
@@ -225,8 +268,10 @@ sc_repl_disable() {
     sc_repl_debug "sc_repl_disable() ... done"
 }
 
-sc_repl_setup() {
-    sc_repl_debug "sc_repl_setup() ..."
+sc_repl_init() {
+    local defaults
+    
+    sc_repl_debug "sc_repl_init() ..."
     sc_version
     if ! sc_repl_asserts; then
         sc_repl_error "ShellCheck REPL startup assertions failed"
@@ -238,17 +283,21 @@ sc_repl_setup() {
     ## SC1090: Can't follow non-constant source. Use a directive to specify
     ##         location.
     ## SC2034: 'var' appears unused. Verify it or export it.
-    ## SC2154: 'var' is referenced but not assigned.
     ## SC2155: Declare and assign separately to avoid masking return values.
     ## SC2164: Use 'cd ... || exit' or 'cd ... || return' in case cd fails.
-    SHELLCHECK_REPL_EXCLUDE=${SHELLCHECK_REPL_EXCLUDE:-1001,1090,2034,2154,2155,2164}
+    defaults=1001,1090,2034,2155,2164
+    sc_repl_debug "- defaults: ${defaults}"
+    SHELLCHECK_REPL_EXCLUDE=${SHELLCHECK_REPL_EXCLUDE:-${defaults}}
     sc_repl_debug "- SHELLCHECK_REPL_EXCLUDE: ${SHELLCHECK_REPL_EXCLUDE}"
     sc_repl_enable
-    sc_repl_debug "sc_repl_setup() ... done"
+    sc_repl_debug "sc_repl_init() ... done"
 }
 
 sc_wiki_url() {
     echo "https://github.com/koalaman/shellcheck/wiki/$1"
 }
 
-${SC_REPL_SETUP:-true} && sc_repl_setup
+if ${SC_REPL_INIT:-true}; then
+    sc_repl_init ""
+fi
+